Split server.py into more manageable files
Originally server.py contained both methods and the server itself, but Uvicorn wouldn't run if the Starlite app wasn't put in the main document. So to fix this and in the interest of keeping things more organized, I renamed server.py to methods.py and created a new server.py file to host strictly the server and its routes.

I prefer it like this as I think it keeps things cleaner, also, so this is how it'll be structured going forward.

Additionally, I need to rework Enums as a whole as they are stored as just strings (i.e. "State.PREGAME") in the database and that is dissatisfactory. Most likely I'll settle for a simpler option.