Having issues with SQLite. Cleaned up code further
I removed some unnecessary values from classes such as the Hand class, updated the ER diagram to be more accurate and modified the database creation code to be more consistent with the diagram. Added some type hints to functions.py and a bit of testing code to server.py.

Additionally, added a simple timeline for what I plan to add next.