Fleshed out classes
Improved the GameSettings class with a lot of customizability to, potentially, enable custom or funny gamemodes in the future and moved the initailize() method to be under it instead of the Game class. Also, cleaned up the code as a whole. Modified the "to_json" decorator to change the __repr__() functions of each class to return a dictionary so that printing a Game object will print it entirely as a dictionary. This allows very easy conversion to JSON.