Implemented "get players in game" option
Added a URL route that returns the verbose names of the players in a game.

This was a funny problem because of how difficult it was without needing to be at all; getting information from session files, apparently, returns a bytes object. Sending a bytes object to a client (in this case, Postman) converts it to base64 encoding. Because of that, I had to reverse engineer why the contents of my session files were stored in base 64 format (they weren't) and had to figure out how to convert it from bytes to string to base 64 decode it (I didn't have to do that).

Fortunately, I got it all sorted when I stepped back a bit and separated things out a bit.

Also changed the join URL from "/api/game/join/{game_id}/" to "/api/game/{game_id}/join" to match the player request URL and keep everything clean.