Fixed logic error
Accidentally only let users join a game NOT in the "pregame" state, which is completely backwards. Fixed that here and also set up expired sessions to be cleared every request. This is almost certainly an issue if we have a large volume of requests but for the time being, it's sufficient to ensure data on disk isn't filled over time.