Slowly starting to implement SSEs
In-progress SSE (server-side event) implementation. My hope is that clients can navigate to "/api/game/{game_id}/notify" and be notified when a change is made in that game, that way they can request for the new board layout.

It's a _little_ less optimized than just sending the board data outright to clients but this is a nice way to not have to avoid about authentication for notifying specifically.