Implemented simple SSE system
Added a system that sends a message to clients connected to "/api/game/{game_id}/notify" when requested. The plan is to now call this whenever a change has been made to the game in question.

I confirmed that notifications are only sent out to individuals listening to that specific game ID URL, so there is no overlap of game notifications.