Added group validation
Groups are sets or runs in Rummikub, so I added validation to the Group class so you can call "group_obj.valid()" to determine if the group is a valid set or run or not. Currently there isn't a distinction made between the two but the way I implemented it, as two separate methods, should make that easy to add if desirable.