Fixed Group validation check
Originally, I didn't account for Jokers at all while checking sequences and colors. This has been fixed.

The only issue that still remains is that validating a group will rearrange any "extra" jokers (jokers that are on the ends) to the ends of a run, wherever there's space, which may reorder what the client sent. I can't think of a good way to resolve this but honestly, it's more of a visual issue than anything else, so I'm not too worried about it for the time being.