Attempting to work out this cookie business
The plugin I'm using, and Godot as a whole, want me to use the HTTPRequest node as just that - I node - but I would like to use it in a singleton so it's not attached to any scene in particular, which makes things significantly more difficult. I'm trying to work out a way to finagle this at the moment but most likely, I'll make a scene that will hold all other scenes and that will hold the HTTPRequest/CookieHTTPRequest node.