Starting to implement SSEs client-side
I've begun work on the client-side portion of SSEs to detect when other players make a move and update the board accordingly. This should allow me to "toggle" the board being updated with the "ping" route to test if SSEs work. Optimally, all users "listening" to a game, at the right URL and with the right permissions, should be notified.